+++ src/toggle_attribute.ts
@@ -6,8 +6,8 @@ export function toggleAttribute<E extends Element>(
   name: BooleanAttributesOf<E>,
   force?: boolean
 ): boolean;
-export function toggleAttribute<E extends Element>(
-  element: E | Maybe<E> | null,
+export function toggleAttribute(
+  element: Element | Maybe<Element> | null,
   name: string,
   force?: boolean
 ): boolean;